# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FurniFlip API - AI-powered furniture photo editing and ad generation
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project and upload original image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        originalImageUrl:
          type: string
        imageCount:
          type: integer
        adCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - originalImageUrl
        - imageCount
        - adCount
        - createdAt
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        originalImageUrl:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/EditedImage"
        ads:
          type: array
          items:
            $ref: "#/components/schemas/Ad"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - originalImageUrl
        - images
        - ads
        - createdAt
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
...
    EditedImage:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [background_removed, staged]
        imageUrl:
          type: string
        roomStyle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - imageUrl
        - createdAt
...
    Ad:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        condition:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - description
        - createdAt
...